fix(cgpt): Fix broken support for full 64-bit attribute field
The command help text for add claims it can set the raw 64-bit attribute
value but then silently casts the argument to a 32-value which is then
explicitly cast into the upper 16-bits of the attribute field. WTF.

Fix this nonsense and use the full 64-bit value everywhere applicable.
Luckily nothing actually depends on the old broken behavior.

+++ cgpt/cgpt.h
@@ -125,7 +125,7 @@ void SetSuccessful(struct drive *drive, int secondary, uint32_t entry_index,
 int GetSuccessful(struct drive *drive, int secondary, uint32_t entry_index);
 
 void SetRaw(struct drive *drive, int secondary, uint32_t entry_index,
-           uint32_t raw);
+           uint64_t raw);
 
 void UpdateAllEntries(struct drive *drive);
